Restructuring folder structure for templates and fixing unit test cases
