+++ templates/sls-app-template-go/functions/function1/config/dev-config.json
