+++ templates/sls-app-template-go/functions/function1/config/prod-config.json
