+++ templates/sls-app-template-go/functions/function1/config/stg-config.json
