+++ templates/sls-app-template-go/functions/function1/main.go
