+++ templates/sls-app-template-go/functions/function2/config/dev-config.json
