+++ templates/sls-app-template-go/functions/function2/config/prod-config.json
