+++ templates/sls-app-template-go/functions/function2/config/stg-config.json
