+++ templates/sls-app-template-go/functions/function2/main.go
