+++ templates/sls-app-template-go/functions/function2/main_test.go
