--- templates/sls-app-template-java/.gitignore
@@ -1,9 +1,0 @@
-# Add any directories, files, or patterns you don't want to be tracked by version control
-.serverless
-node_modules
-.DS_Store
-.classpath
-.project
-.settings/
-target/
-.vscode/
