+++ templates/sls-app-template-java/src/main/java/com/slf/services/functions/function1/Function1.java
