+++ templates/sls-app-template-java/src/main/java/com/slf/services/functions/function2/Function2.java
