+++ templates/sls-app-template-java/src/main/resources/functions/function1/dev.properties
