+++ templates/sls-app-template-java/src/main/resources/functions/function2/stg.properties
