+++ templates/sls-app-template-java/src/test/java/com/slf/services/functions/function2/Function2Test.java
@@ -9,23 +9,23 @@ import com.slf.stubs.ContextStub;
 import java.util.HashMap;
 import java.util.Map;
 
-public class Function1Test {
+public class Function2Test {
 
 
-	private Function1 f1 = new Function1();
+	private Function2 f2 = new Function2();
 
-	public  Context context = new ContextStub("Function1-dev");
+	public  Context context = new ContextStub("mydomain80t-myname80t-function2-prod");
 
 	@Test
 	public void shouldExecuteRequest() {
 		Map<String, Object> input = new HashMap();
-		input.put("key", "value");
-		Response response = f1.handleRequest(input, context);
+		input.put("key", "value_2");
+		Response response = f2.handleRequest(input, context);
 		Map<String, String> output = (Map) response.getData();
 
-		assertTrue(output.get("name").equals("value"));
+		assertTrue(output.get("name").equals("value_2"));
 		// Should have read the value from the config
-		assertTrue(output.get("config_key").equals("config_value_dev"));
+		assertTrue(output.get("config_key").equals("config_value_prod_2"));
 
 	}
 }
