+++ templates/sls-app-template-nodejs/functions/function1/config/dev-config.json
