+++ templates/sls-app-template-nodejs/functions/function1/config/prod-config.json
