+++ templates/sls-app-template-nodejs/functions/function1/config/stg-config.json
