+++ templates/sls-app-template-nodejs/functions/function1/test/test.js
