+++ templates/sls-app-template-nodejs/functions/function2/config/dev-config.json
