+++ templates/sls-app-template-nodejs/functions/function2/config/prod-config.json
