+++ templates/sls-app-template-nodejs/functions/function2/config/stg-config.json
