+++ templates/sls-app-template-nodejs/functions/function2/test/test.js
