--- templates/sls-app-template-python/.gitignore
@@ -1,15 +1,0 @@
-# Add any directories, files, or patterns you don't want to be tracked by version control
-.serverless
-node_modules
-.DS_Store
-.classpath
-.project
-.settings/
-target/
-.vscode/
-__pycache__/
-*.py[cod]
-*$py.class
-.cache
--venv
--library
