+++ templates/sls-app-template-python/functions/function1/config/dev-config.ini
