+++ templates/sls-app-template-python/functions/function1/config/stg-config.ini
