+++ templates/sls-app-template-python/functions/function2/config/dev-config.ini
