+++ templates/sls-app-template-python/functions/function2/config/stg-config.ini
