+++ templates/sls-app-template-python/functions/function2/main.py
