+++ templates/sls-app-template-python/functions/function2/test/test_hello.py
@@ -1,4 +1,10 @@
+import sys
+import os
 import json
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+
 import main
